Fix test and RAM slot power
